Post buck events for artifact compression and decompression
Summary:
Sometimes we spend a significant portion of time compressing or
decompressing an artifact, but this was never logged to the event
bus, so we'd never seen information about this in traces or the
super console.

This diff adds new compression and decompression events for these
stages so they become more visible.

Test Plan:
Ran build.  Saw `artifact_compress` and `artifact_decompress` steps
in super console.