Transform path arguments under cygwin
Summary:
When running in a cygwin environment, `buck build buck` will invoke
python with DOS-style absolute paths, but the python interpreter will
treat it as a relative path. These paths need to be transformed to
cygwin paths in order to be used in the cygwin python environment.

Test Plan: Run buck build buck and see DEFS get included properly.