Fix Xcode asset catalog compiler command going over kernel length limit
Summary:
When running the asset compiler on a project with hundreds or
thousands of `.xcassets` bundles, we could go over the kernel limit
on the length of a command line.

We actually hit this in a project:

    Foo.build/Script-E1F174220000000000000004.sh: line 3: path/to/compile_asset_catalogs_build_phase.sh: Argument list too long

This fixes the issue by using the `fromfile_prefix_chars` feature of Python
`argparse` to allow writing the `Foo.xcassets` paths to a temp file and then passing
them to the script using `@filename`:

  path/to/compile_asset_catalogs_build_phase.sh @/tmp/args

Test Plan:
Added unit test. Confirmed it failed with `ant java-test -Dtest.class=NewNativeTargetProjectMutatorTest`.

Applied patch, re-ran `ant java-test -Dtest.class=NewNativeTargetProjectMutatorTest`, confirmed it passed.