Avoid spamming cache client connections
Summary:
The default size for pooled connections is `5`, which means we easily
overflow this number with typical default `-j` values.  When this
happens, even if just by one, we start spamming connection creation
and causing way too many connections to the cache server.  This diff
sets the appropriate config to a large value to avoid this.

Long-term we should witch to a better HTTP client, something like
OkHttp, which will allow us to handle this stuff explicitly.

Test Plan:
Ran a build with and without this and used netstat to verify that
HTTP connections were in fact being reused (a max of 10 in a build)
when set (as opposed to something like 130 without).