Adds a new `--include-tests` flag to `audit dependencies` to print tests
Summary: One of the feature requests I heard after I told people about the `audit dependencies` command was to provide a way to get all of the test targets for all of the dependencies. This diff adds that feature via the `--include-tests` flag. When provided, it will print out test targets as if they were dependencies. Without the `--transitive` flag, it will print only tests for that exact library, and with the `--transitive` flag it will print out all tests and the tests dependencies as well.

Test Plan: Integration Tests.