Allow arbitrary macros to write their output to a file
Summary:
Allows any other macro to output its output to a file. The path to this file
will be output as the output with "@" prepended. This means it can be passed
straight to apps like javac and Python's optparse.

Usage is to prepend the normal macro name with an "@" symbol.

eg: "foo $(@classpath //some:target)" expands to "foo @/temp/directory/file"

Test Plan: buck test --all