Make `--json` flag of `audit tests` return a map instead of a list
Summary: Since `--json` is normally used for things like automation is makes more sense to give back as much information as possible. This also has the benefit of matching the current functionality of `audit input`, which prints out a list without the `--json` flag but prints a map with it.

Test Plan: Integration Tests.