Fix escaping for gen_rule on Windows
Summary:
This change has `ExecutableMacroExpander` use platform-aware
escaping instead of always using bash escaping.  This is fine
to do because we always use `cmd.EXE` on Windows, and `bash`
on everything else.

This also updates what we log to use the platform-aware escaping
so a developer can run the command that failed more easily (I
say more because we still escape for bash for the environment
variables and the changing directory command, but this is still
strictly better than what we had before even if it isn't 100%
correct).

Test Plan:
* `ant test -D"test.class"="ExecutableMacroExpanderTest"`
* `.\bin\buck build buck-win` gets a lot farther on Windows