TargetsCommandOptions extends AbstractCommandOptions instead of BuildCommandOptions.
Summary:
This cleans up an inappropriate inheritance hierarchy. This appears to have been
used primarily as a convenience to get access to `getArgumentsFormattedAsBuildTargets()`,
which is a bad reason for subclassing. Other options classes, such as
`AuditCommandOptions`, already have their own implementation of
`getArgumentsFormattedAsBuildTargets()`, so it seemed reasonable to follow suit here as
a stopgap solution.

I also tried to make `getArguments()` and `getArgumentsFormattedAsBuildTargets()` methods
of `AbstractCommandOptions`, but that was problematic because `RunCommandOptions` has a
funky implementation with respect to `getArguments()` and its use of
`ConsumeAllOptionsHandler`, so moving `getArguments()` to the superclass was not a
straightforward solution.

Test Plan:
Previously, running `buck targets --help` would list options specific to building,
such as `--keep-going` and `--build-report`, which are irrelevant for `buck targets`.
Now those no longer appear.