Expose tests field from AbstractCxxLibrary
Summary:
We want test rules to be able to directly include both the
public and private headers of any rules they test.

To do this, we need to ask a library what nodes test it.

This implements that check by exposing the `tests` field of
`CxxLibraryDescription` in `AbstractCxxLibrary` and `CxxLibrary` so we
can access it from `CxxLibraryDescription`.

Test Plan: `ant java-test -Dtest.class=CxxLibraryDescriptionTest`