package-private is sufficient for @VisibleForTesting: should not be protected unless overridden.
Summary:
I noticed a handful of methods that were declared `protected` in classes that
have no subclasses (and do not appear to be designed to be subclassed).
I think this was an errant assumption that this was necessary because these
were `@VisibleForTesting`.

Check your Java visibility chart, folks!
http://docs.oracle.com/javase/tutorial/java/javaOO/accesscontrol.html

`protected` is also package-private, but package-private is all we need
for testing. Also, marking things as `protected` is confusing because it
suggests that (1) there are subclasses and (2) such members are designed
to be overridden.

If there are other areas of the code that also have this issue,
please clean them up if you are aware of them.

Test Plan: CI