Fix linking for iOS Simulator
Summary:
We were unable to link binaries when targeting the iOS
simulator, because I incorrectly assumed based on the documentation
that both iOS and iOS Simulator were targeted with the linker flag
`-mios-version-min=...`.

In fact, when linking binaries for iOS simulator, you have to
specifically pass the undocumented flag
`-mios-simulator-version-min=...` or the linker warns:

  ld: warning: ObjC object file (buck-out/bin/path/to/lib#compile-pic-src_Foo_Bar.m.o,iphonesimulator-x86_64/src/Foo/Bar.m.o) was compiled for iOS Simulator, but linking for MacOSX

This fixes the warning and adds a test.

While I was at it, I discovered we were setting the C++ linker
flags but not the C linker flags.

Test Plan: Unit test updated. `ant java-test -Dtest.class=AppleCxxPlatformsTest`