Big Buck perf fix [1/2]: Add Xcode version to AppleSdk
Summary:
@agallagher and I discovered today that Apple builds
are extremely slow, because `AppleCxxPlatform` uses `HashedFileTool`
for `clang` and other tools.

This causes `RuleKey` generation to repeatedly hash the `clang` binary
over and over. This binary is 50 MiB or more in size, which makes
builds very slow.

We don't cache it because we can only cache hashes for files under
Watchman (or else the caches would get stale), which doesn't include
Xcode utilities.

This diff is part 1 of 2 of the fix. We want to move from `HashedFileTool`
to `VersionedTool`, but we need to include in the version cache key
both the version of Xcode as well as the SDK name and version.

In this diff, I expose the version of Xcode in `AppleSdk` and add tests.

Test Plan:
Tests added. `ant java-test -Dtest.class=AppleSdkDiscoveryTest`

Ran buck build //path/to:lib in Apple repository, confirmed in logs that
Xcode version was detected.