Adds ability to generate Xcode workspaces from `apple_binary` and `apple_library` targets
Summary:
One of the common pain points we have is that building and running tests takes FOREVER. This is due to the sheer number of tests that get included on any of our big projects. To solve this, many people on our team have gotten into the habit of making `xcode_workspace_config` targets for every library for which they run tests. This ensures that only the tests they need are included in the workspace and generally leads to a much quicker workflow. These workspaces didn't have much to them, and normally look like this:

```
xcode_workspace_config(
  name='workspace',
  src_target=':Library',
)
```

There is no reason we need to add those 4 lines to every buck file we use, especially when the only thing that changes is the `src_target` parameter. To fix this, this diff allows you to run `buck project ` on `apple_binary` and `apple_library` targets directly. Doing so is equivalent to running it on an `xcode_workspace_config` exactly like above, but with the `src_target` set to whatever `buck project`'s parameter was. This should make it easier to work in the tiny library-specific workspaces without needing to commit these changes to version control.

Test Plan: Integration Tests