Fix race-condition handling classes.jar
Summary:
We currently use an `OutputOnlyBuildRule` to wrap the `classes.jar`
output of the `UnzipAar` rule, so that consumers can reference the
`classes.jar` via a `BuildTargetSourcePath`.  However, this means that
two build rules will now write to the same path (for `classes.jar`)
when they pull from cache.

This appears to be able to cause corruption in how the `CachingBuildEngine`
works, since as soon as the `UnzipAar` rule finishes, it then kicks off
the `OutputOnlyBuildRule` that wraps the `classes.jar` file while uploading
to the cache in parallel.  If the `OutputOnlyBuildRule` fetches it's cached
`classes.jar` file from cache it can possible overwrite the one we're uploading
to cache causing corruption.

This has been this way for a while.  However, until D1834999 went in,
we never had an actual dep on the `OutputOnlyBuildRule`, which meant
it never actually ran.

This diff replaces the uses of `OutputOnlyBuildRule` with a `BuildTargetSourcePath`
using an explicit path to avoid this issue.

Test Plan: `buck test --all`