Properly escape arguments before handing off to `ProcessBuilder`
Summary:
On Windows, `ProcessBuilder` hands whatever we provide as arguments
over to `CreateProcess`.  However, that assumes we've done some
escaping before handing it arguments.  This change makes
`ProcessExecutor` escape the arguments as appropriate prior to
giving `ProcessBuilder` on Windows.  It also renames
`WindowsCommandLineEscape` to be `WindowsCreateProcessEscape` to
better represent what it does and to avoid potential confusion in
the future.

The upside here is that `.\bin\buck test --all` now actually runs on
Windows, so we can have a good idea of what's broken (surprisingly,
not much).

There are two items I need to do as a follow-up to this:
* Fix `AbstractGenRule` to use proper escaping for `cmd.EXE` (it is currently wrong)
* Fix all uses of `ProcessBuilder` by converting them to `ProcessExecutor` (or escaping properly)

Test Plan:
* `.\bin\buck build buck-win`
* `ant clean default`
* `ant test -D"test.class"="WindowsCreateProcessEscapeTest"
* `.\bin\buck test --all` actually runs Java tests
* `ant lint`