Fix perf_test
Summary:
We changed the rule key logging to be a warning which padded
the "level" segment of the log message with a space.  This caused
perf test to stop regex'ing it correctly.

This diff reverts back to verbose logging (since the original change
appears unintentional) and fixes the perf test to be more robust at
handling the log level.

Test Plan: CI