Use the filtered environment in `FakeRepositoryFactory`
Summary:
`FakeRepositoryFactory` was just copying the system environment,
but on Windows, this is non-ideal because the case of the
environment variables doesn't matter.  We already solved this
problem in `EnvironmentFilter`, so just use that to get a sensible
environment.

This also moves `EnvironmentFilter` out of `com.facebook.buck.cli`,
because having sensible deps is nice.

Test Plan:
* No longer saw issues related to the my `PATH` variable being `Path`
* `.\bin\buck build buck-win`
* `ant lint`