Exclude generated classes and inner classes from code coverage
Summary:
The script `assert_code_coverage.py` had a heuristic to ignore
classes whose name start with `Immutable`.

This won't work when we start using `@NewBuckStyleImmutable`, since the
generated classes for `AbstractFoo` are named `Foo` with this style.

This updates the heuristic to look inside the `buck-out/gen` directory
for any generated files and build a set of files inside. We use that
set to ignore jacoco reports on code coverage for generated classes or
inner classes.

Test Plan:
  % buck test --code-coverage --code-coverage-format xml --no-results-cache //test/com/facebook/buck/apple:apple
  % python scripts/assert_code_coverage.py