Modify Eclipse project to workaround issues with Immutables 2.0.
Summary:
Made the following changes to the project:

* Added `build/immutables` and `build/immutables-tests` as source folders.
* Disabled annotation processing.

Now Eclipse users must be sure to run `ant generate-immutables generate-test-immutables` whenever
the source of an immutable needs to be changed. Previously, I was able
to rely on Eclipse's ability to generate the appropriate sources, as it
knew when to run the annotation processor.

Test Plan: Project now builds cleanly in Eclipse.