Replace the heuristic-based CompilationDatabase with the legit CxxCompilationDatabase.
Summary:
This removes `com.facebook.buck.apple.CompilationDatabase` from the codebase,
and any references and tests, in favor of `com.facebook.buck.cxx.CxxCompilationDatabase`.

Because `AppleLibraryDescription` and `AppleBinaryDescription` have calls to
`delegate.createBuildRule()`, which already checks for the `#compilation-database`
flavor and returns a `CxxCompilationDatabase` rule, if appropriate, the changes to the
Apple rules are fairly minimal.

Test Plan:
Updated `CompilationDatabaseIntegrationTest` to verify the new flags.
`buck test --all` succeeds locally on my OS X laptop.
I'm not sure if `CompilationDatabaseIntegrationTest` can be expected to work on Linux.