Add a nonstandard `args` field to entries in the compilation database.
Summary:
In practice, we currently spend a lot of energy making sure that
we escape and unescape a list of args correctly. This seems silly because:

(1) Buck has the clang args as a list.
(2) libclang wants to consume the args as a list.
(3) A list of strings can be represented easily in JSON.

This revision changes things so that we still provide the standard `command`
property in each entry (http://clang.llvm.org/docs/JSONCompilationDatabase.html),
which is a string, but we also introduce the `args` property which is the
array of args that are escaped and joined to produce the `command`.

Perhaps we can lobby for this to be part of the official format.

Moved `CxxCompilationDatabase.JsonSerializableDatabaseEntry` into its own class,
`CxxCompilationDatabaseEntry`, that contains its own deserialization logic that can be
reused across integration tests.

Test Plan:
Updated the following tests:
* `com.facebook.buck.cxx.CxxCompilationDatabaseTest`
* `com.facebook.buck.cxx.CxxCompilationDatabaseIntegrationTest`
* `com.facebook.buck.apple.CompilationDatabaseIntegrationTest`