Model client-side cache errors via `CacheResult`
Summary:
Previously, when buck failed fetching from cache we'd log a message
to the buck log then report a miss.  This makes it very difficult to
aggregate information client errors in any meaningful way.

This diff adds the concept of a cache failure result, which can be consumed
and aggregated via the event bus.

Test Plan: `buck test --all`