Make AbstractAppleBundleDestination implement RuleKeyAppendable
Summary:
This is a follow-up to my diff
`Make AbstractAppleBundleDestination implement RuleKeyAppendable`.

Previously, we could not include `AppleBundleDestination` types inside
`@RuleKeyAppendable` containers.

This makes the type implement `@RuleKeyAppendable` so we can do so.

Test Plan:
Not sure how to test this in an automated fashion
without duplicating most of `DefaultRuleKeyBuilderFactoryTest` --
@simons, any ideas?