Use `ExecutableFinder` to locate the NDK `VersionedTool`s
Summary:
On Windows, we weren't finding the right tool because we were not
looking for `.EXE` files.  Note that tests still don't pass, but at
least we are now trying to use the right tool.

Notably, this makes finding the tool more eager, so if you have a
mis-configured environment, you'll fail early.

Test Plan:
Fired up `NdkLibraryIntegrationTest` in the debugger and verified
that the path we were trying to find actually was the tool on my
machine.