Don't allow PathHashing to hash empty path
Summary:
While developing my diff `Copy resources of test dependencies into AppleBundle rule`
I accidentally introduced a `@RuleKeyAppendable` which contained an empty `Path`.

This was pretty disastrous, as it caused `PathHashing` to recurse
through every file and directory in the project filesystem.

This diff simply adds a `Preconditions` check to ensure passing
an empty `Path` to `PathHashing` is treated as a programmer error (there's
no situation in which this would be valid).

Test Plan: buck test --all