Add a new buck command: buck server status.
Summary:
Prints usage if a subcommand is not specified, or an invalid subcommand is
used.

Currently only supports `buck server status --http-port` ( and `--json`).

There is a ton of copypasta from `AuditCommandRunner`. I plan to fix that in
subsequent diffs.

Test Plan:
Tested the following commands (with and without buckd running):
```
$ buck server status --http-port
http.port=-1
$ buck server status --http-port --json
{"http.port":"-1"}
```