Don't throw exception when [cxx] default_platform not found
Summary:
In my previous diff `Allow specifying a default C++ platform in .buckconfig`,
I originally threw an exception if the `default_platform` specified in config
was not found.

This was a little harsh for build infrastructure which doesn't actually
depend on compiling C++; if the default platform didn't exist, Buck would
throw even if it didn't need to compile anything.

In this diff, I:

1) Move the code to choose the default platform out of `KnownBuildRuleTypes`
and into `CxxPlatforms`

2) Loosen the exception and turn it into a warning

3) Add tests

Test Plan:
New unit tests added.
  ant java-test -Dtest.class=CxxPlatformsTest
  ant java-test -Dtest.class=KnownBuildRuleTypesTest