Delete output path for AppleBundle before trying to write it.
Summary:
I saw an error where Buck complained that it could not overwrite
the `.zip` file. Because a `ZipStep` claims it can either create
or update a ZIP archive, it seems prudent to force delete the
destination path before trying to write it.

I'm a little surprised that I had to do this because I thought
we did this automatically for `BuildRule#getPathToOutputFile()`,
but perhaps we don't because that screws up ABI rule calculation
in certain cases?

Test Plan: Was able to build my `apple_bundle()` again.