Call `outputTrace` on listeners even on fatal builds
Summary:
Previously, when buck would die via `ctrl-c` we would not process the
build listeners, meaning we'd lose any special logging for that
particular build.  This diff moves the build listener tracing to the
`finally` clause so it always happens.

Test Plan:
Added unittest that kills the build and verifies that build listeners
are still processed.