Fix apple_test build when no platform specified
Summary:
Previously, when running:

  buck build //path/to:apple_test

Buck would crash with an NPE due to trying to resolve the output
of a rule with no output.

The core issue was that `AppleTestDescription` needed similar logic
to handle the unspecified default platform as the logic in `CxxBinaryDescription`
and `CxxLibraryDescription`.

Until we unify the rules, this adds explicit handling for "no platform specified"
to `AppleTestDescription`.

Test Plan:
New integration test included. `buck build //path/to:apple_test` in
Apple repo.