Support Java tests with JNI dependencies
Summary:
Allow JUnit/Robolectric to be used for integration tests that depend on
native code (as long as the native code can be compiled without Android
headers).

JavaTest now spiders its dependencies to create default/shared
cxx_library rules the same way that JavaBinary does.  Those are added to
a symlink tree, which is added to the java.library.path.

This is optional on a per-build-rule basis, defaulting to off.

Test Plan:
Ran "buck test --all" in the test project.
Ran the top-down stress test in the test project.
I'll make it a real integration test once we deal with jni.h