Change CommandRunner implementations to get params in the run method instead of the constructor.
Summary:
CommandRunnerParams contains the union of all things all commands need to run. This is currently
passed into the command runners' constructors, which means we cannot change any of those in
response to command line flags without parsing the command line flags separately just for that,
as is the case for verbosity. This changes the command runners to get the params in the run
method instead of the constructor. In this way we can parse command line flags that affect
objects in CommandRunnerParams with the AbstractCommandOptions subclasses.

Test Plan: CI