Change AbstractCommandOptions to get BuckConfig as parameter to methods instead of constructor.
Summary:
AbstractCommandOptions gets a BuckConfig in its constructor so it can use it later to marshal
command line arguments. This means that we BuckConfig has to be created before
AbstractCommandOptions, so it cannot have fields be overridden easily from the command line.
This removes BuckConfig as a prerequisite to creating AbstractCommandOptions.

Test Plan: CI