Natively run Apple tests [1/3]: Add linking as Mach-O Bundle
Summary:
This is part 1/3 of natively running Apple tests in Buck.

To run Apple "application" tests, we need to link Apple tests as a
Mach-O bundle which is injected into a test host application.

This is the groundwork for adding that functionality.

Test Plan: Unit test included. `ant java-test -Dtest.class=CxxLinkableEnhancerTest`