New test util method ProjectWorkspace.copyRecursively()
Summary:
I'm writing a few integration tests for the `xctool` utility,
which consists of a tree of files with a binary, some libraries, etc.

I don't want to check in multiple copies of this tool to the test
repo, so this diff introduces a new method `ProjectWorkspace.copyRecursively()`
which copies one testdata directory to the temporary project workspace.

Test Plan: Integration test included. `ant java-test -Dtest.class=ProjectWorkspaceTest`