Upgrade Immutables.org to 2.0.11-SNAPSHOT to fix IntelliJ build
Summary:
This upgrades Immutables.org to upstream rev
6edec0106f571916f9221c3dc3ab657dd317c87f (2.0.11-SNAPSHOT) to fix
the IntelliJ build.

There are still issues with the Eclipse build, but this will improve life for the Buck developers who use IntellIJ (they no longer have to run `ant ide` to prepare the build).

Test Plan:
Built in IntellJ 14. Confirmed immutable classes
were created in `build-ij/immutables` and that completion
worked in the UI.