Add checks for forbidden Java APIs to ant build
Summary:
There are a number of Java APIs which take an implicit
`Charset` or `Locale`. These are not suitable for use by Buck, since
the default `Charset` and `Locale` are not under our control.

This adds an open-source tool `forbiddenapis` (Apache 2.0 license)
which runs on our `.class` files and raises errors for any APIs.

It comes with a default set of forbidden APIs, and we can extend
this set with whatever APIs we want to ban by adding them to
a file (I chose `config/forbiddenapis.txt`).

To invoke the checker on non-test classes:

  ant forbidden-checks

And on test classes:

  ant forbidden-checks-tests

I didn't integrate this with the default build because we have a lot
of forbidden APIs to fix:

https://gist.github.com/bgertzfield/af29ab7c9a9f7df05da1
https://gist.github.com/bgertzfield/a06715cf739ad9731f20

Test Plan: `ant forbidden-checks forbidden-checks-tests`