Ensure ZipStep maintains Posix file permissions
Summary:
When adding support for `apple_test` rules with
application-hosted tests, I found that `ZipStep` creates
archives which ignore Posix permissions.

This wouldn't be a big deal, except that for Apple app tests,
we use the output of an `apple_bundle` rule, which contains
a zip archive of a directory including executable files.

When unzipping the resulting zip, the binary wouldn't be
executable, and the test harness would fail to run.

This adds logic to `ZipStep` to maintain Posix file
permissions when possible, along with an integration test.

Test Plan: `ant java-test -Dtest.class=ZipStepTest`