CxxLinkableEnhancer: Ensure Mach-O bundle loader is added to deps of link rule
Summary:
In my diff `Natively run Apple tests [1/3]: Add linking as Mach-O Bundle`,
I forgot to add the passed-in bundle loader to the dependencies of the rule
being linked.

If the bundle loader was itself the output of another rule, this would
manifest itself as occasional build failures (if the linkable was
built before the bundle loader finished building).

This adds the bundle loader to the deps of the newly-created link rule
and includes a test.

Test Plan: Unit test included. `ant java-test -Dtest.class=CxxLinkableEnhancerTest`