Add suspend/resume events for build rules
Summary:
When switching to a top-down build engine, we'll be processing a given
build rule in two different parts: initially in the walk down the action
graph while fetching from cache and again in th walk back up when actually
building.  Since these two stages aren't necessarily executed on the same
thread, or contiguous, we can't model them with just "started" and "finished"
events.

This diff adds two new `BuildRuleEvent`s to model this, `Suspended` and
`Resumed`, and updates the various listeners to handle them.  In particular,
the super console will remember the times from previous suspended segments
of a build rules execution so it can resume from this amount when resuming
the rule.

Test Plan: Added unittests for listeners with non-trivial changes.