[new IJ project gen] Drop support for ProjectConfig.
Summary:
This removes references to ProjectConfig targets from the new project
generator. These were initially only used to resolve conflicts but
that's only necessary if we wanted to build with IntelliJ which is
not something the new project generator supports yet.
We still use ProjectConfig to figure out which targets to include
when running `buck project`, removing that will come later.

Test Plan: buck test