Easy: New test utility class FakeBuildEngine
Summary:
I'm adding a feature to `TestRunning` to run certain
tests separate from the rest of the tests.

To test that feature, I want to schedule a bunch of `Step`s and
make sure they're executed in a certain order.

This diff introduces a test utility class `FakeBuildEngine` which I'll
use in the tests for `TestRunning`.

Test Plan: Unit test included. `ant java-test -Dtest.class=FakeBuildEngineTest`