Add a PowerShell script to launch soyweb locally
Summary:
This is nearly the same as `docs/soyweb-local.sh`, but works on Windows.

I never thought I'd have to look at the binary encoding of a JSON file,
but for this diff I had to to figure out that soyweb can't processes a
Unicode-encoded JSON file.

Test Plan: launched it and loaded it in my browser