super console: update cache stats on artifact finished events
Summary:
Previously, we'd collect cache stats via the build rule finished
events.  However, we get this information on the bus a lot sooner
in the form of an artifact finished event, meaning we can display
this before the rule has finished being processed.

This is particularly useful in top-down building, since we may not
finish processing the rule until long after doing the initial fetch.

Test Plan: Performed a top-down build and saw cache stats on downwards pass.