BuildTargetSourcePaths with different paths must not be equal
Summary:
Previously, a `BuildTargetSourcePath` just based equality on it's
`BuildTarget`.  That meant the two `BuildTargetSourcePath` which referred
to different outputs of the same rule were equal and got collapsed
in sets.

This diff makes includes the optional output path in the `asReference`
object which is used in the implementing a source paths equality and
hash code.

Test Plan: added unittest