New class TeeInputStream
Summary:
I need to take an `InputStream` from a launched `Process`
and simultaneously read its data as well as write that data to
a file.

This introduces a new class `TeeInputStream` modeled after the Unix
`tee` command which takes an `InputStream` and an `OutputStream` as
parameters. Whenever `TeeInputStream.read()` is called, it reads
the data from the `InputStream`, writes it to the `OutputStream`,
and returns it.

Test Plan: Unit tests included. `ant java-test -Dtest.class=TeeInputStreamTest`