Support concept of "runtime" dependencies
Summary:
Some rules rely on build products from (transitive) build dependencies
to be available on disk at runtime (e.g. Java tests).  When we move to
a top-down build engine, we can't rely on normal build dependencies.
This diff adds a new `HasRuntimeDeps` interface which rules can use to
signal to the build engine dependencies that need to always be emerged
locally during a build and uses this in `JavaTest`.

Test Plan: added test for `JavaTest`