Use `EventBus` instead of `AsyncEventBus`
Summary:
Even though we give it a single thread executor, an `AsyncEventBus`
makes no guarantees about processing the handlers in order and it's
possible, albeit rare, for it to execute the events out-of-order even
on a single thread.

This diff switches to using a `EventBus` called using the executor
service in `BuckEventBus` to guarantee that subscribers are run in-order.

Test Plan: `buck test --all`