Run ant with same heap size as Buck itself
Summary:
We've seen some systems which fail to build Buck with
an OOM from ant while generating immutables:

https://gist.github.com/bgertzfield/4ce6deb6435579396a70

This bumps the max heap size for ant to be the same as Buck itself.

Test Plan:
Added log message, confirmed `ANT_OPTS` was set in
environment when running `buck build buck` on a clean tree.