[apple test] Fix OS X logic and app tests
Summary:
I had a `TODO` in `AppleTestDescription` to stop hard-coding
`iphonesimulator` as the SDK.

When I fixed this, I found a bug in `xctool` where it threw an exception
when trying to chdir to a nil directory when running OS X logic or app
tests directly from a built binary.

I fixed the bug and sent a pull request upstream which I merged:

https://github.com/facebook/xctool/pull/519

This updates our local copy of the binary to include the fix and
adds integration tests to build and run OS X logic and app test.

Test Plan: Integration tests included. `ant java-test -Dtest.class=AppleTestIntegrationTest`