NdkCxxPlatforms: support building with clang and libc++
Summary:
This diff teaches the `NdkCxxPlatforms` code how to set the necessary
tools and paths to build C/C++ libraries with clang and libc++.
Most of the work in this diff is making the platform building process
more configurable and so adds some enums and classes to represent these
new concepts.

I have not put a lot of thought into the UI.  This diff just adds four new
`.buckconfig` parameters as a bare minimum to allow users to start using
it: `compiler`, `gcc_version`, `clang_version`, `cxx_runtime`.  I wasn't
able to collapse the `gcc_version` and `clang_version` items into a single
`version` parameter, as even when building with clang, the NDK needs to
find utilities/tools in the GCC toolchain (e.g. `objcopy`, `ar`, `ld.gold`).

Test Plan:
Added integration test that builds C/C++ for the core combinations or
compiler, runtime, and architecture that we want to support and verifies
debug paths are sanitized correctly.