Move out android specific logic into AndroidReactNativeLibrary.
Summary:
Since packaging the JS bundle and images/assets is specific to android, it
makes sense to create a new rule that implements `AndroidPackageable`.
Now the sole responsibility of `ReactNativeBundle` is to call into the JS
packager.

Another motivation for this is a subsequent diff that creates an android_resource()
rule which needs to depend on `ReactNativeBundle`, and so
`AndroidReactNativeLibraryDescription` would end up returning an android_resource
rule, which seemed weird.

Test Plan: CI