ArtifactCache: allow storing an artifact via multiple keys
Summary:
We're moving to a model where we'll use multiple styles of rule keys
to describe a single rule.  This diff, therefore, allows indexing the
output of a single rule into the cache by multiple different rule keys,
allowing lookups with any of the passed in keys.

Currently, this is supported naively in the various implementations:
a multi-key set is converted into a single set for each key.

Test Plan: Added multi-key store tests for implementations.