Fix AndroidReactNativeLibraryIntegrationTest.testEditingImageRebuildsAndroidResource().
Summary:
We do not list all inputs of `android_react_native_library` as sources of the
android resource rule, because we don't want the rule to be re-built if any
unreferenced input file changes. However, that breaks the integration test.

This diff addresses both the issues by passing the hashes of the actual inputs
to the android resource rule.

Test Plan: CI