Make installApk handle Exopackages
Summary:
Instead of relying on on separate logic to install Exopackages, make the
installApk function of the AdbHelper check if the package is an
Exopackage, and call the ExopackageInstaller directly from there.  When
adding new call sites for installApk, it is no longer necessary to
duplicate the logic for installing Exopackages.

Test plan: buck test