Do not format ConsoleEvent message if there are no arguments
Summary:
There are cases where the message passed to `ConsoleEvent` contains character
sequences that are interpreted as format arguments for `String.format()`. This diff
"fixes" the issue by never calling `String.format()` in the first place if there are no
arguments. Assuming that if you are passing argument you probably control the
message string as well.

Test Plan: buck test