Don't download artifacts if there's a newer version of it locally
Summary:
The current maven importer blindly downloads whichever version of a
dependency it is calculated is required. This means it's possible
to replace a newer version of a library already in the local
repository with one that's older, which is obviously suboptimal.
Fortunately, Maven uses a mechanism for encoding version numbers in
downloaded artifacts, so we can detect the current version of a dep
fairly easily (the assumption being that the dep needs to be
updated if we can't tell the version number)

Test Plan:
  buck test --filter ResolverIntegrationTest test/com/facebook/buck/maven