BuildInfoRecorder: add concept of "build" metadata
Summary:
This adds the concept of "build" metadata to the `BuildInfoRecorder`
so that the build engine can attach non-stable metadata to the artifact.
The "build" metadata gets passed to the cache via the metadata parameter
in the caching API, so that it doesn't affect the actual output of the
build rule (so it can be effectively deduplicated by the caching backend).

Test Plan: Updated tests to test for build metadata methods.