Strip down bootstrapper classpath in Ant build, add warning comments
Summary:
The whole point of the bootstrapper is to keep the system ClassLoader
free of classes that might conflict with ones used by plugins running
inside the in-process compiler. The Ant build was building it with
the default classpath from the `<ecj>` macro, which has a bunch of stuff
in it. No bad dependencies had crept in, but better safe than sorry.

Also added some comments to both the Ant and Buck builds to help
prevent bad things happening in the future.

Test Plan: Run the Ant build, verify it still succeeds