Scrub ZIP timestamps from android build products
Summary:
A couple of the android build steps delegate to third-party tools
to construct ZIPs.  These ZIPs contain timestamps which make our
build output non-deterministic.  This diff adds a `ZipScrubberStep`
which scrubs timestamps in the local and central directory headers
of a ZIP file in-place and uses it to make APK builds deterministic.

Test Plan: added tests