Infer empty "src" parameter from "name" for export file rules.
Summary:
If the user omits the "src" parameter, the
ExportFileDescription can infer this from the "name" field.
However, the logic in `TargetNode.getInputs` cannot, as it
scans the description itself without any knowledge of how to
interpret the fields other than the types used.

Added a check for ExportFile rules. If there are no inputs detected,
then use the name as an input.

Test Plan:
Added integration test with an export file rule that contains "name" but not
"src". Checked that buck audit input now infers the name as an input.