Change category of annotation processor events
Summary:
Previously these were traced with a category of `buck`. However, really they're
not part of Buck, and reporting them in their own category allows one to search
for the category name and obtain a summary of all events in that category.

Test Plan:
Do a build with an annotation processor that uses Buck's Tracing API
In the resulting `build.trace`, search for the name of the processor
Observe that all the annotation processor's custom events are summarized at the
bottom, alongside the events Buck outputs for the top-level methods in the
processor.